Moved Instruction Components to separate file
